Reorganize package
(also, add `y` and `n` as possible values for BoolSetting)
