+++ coveo-settings/coveo_settings/__init__.py
@@ -1,0 +1,8 @@
+from coveo_settings.any_setting import AnySetting
+from coveo_settings.bool_setting import BoolSetting
+from coveo_settings.dict_setting import DictSetting
+from coveo_settings.float_setting import FloatSetting
+from coveo_settings.int_setting import IntSetting
+from coveo_settings.path_setting import PathSetting
+from coveo_settings.setting_abc import Setting
+from coveo_settings.string_setting import StringSetting
